+++ Packages/webxr/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for Spectator Camera.
 
+### Fixed
+- WebXR input profiles loading on SteamVR.
+
 ## [0.15.0] - 2022-06-04
 ### Added
 - GetButtonTouched to WebXRController, to support buttons touched values.